+++ elphick/mass_composition/__init__.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from .mass_composition import MassComposition
 from .stream import Stream
+from .flowsheet import Flowsheet
 from .mc_xarray import MassCompositionAccessor
 
 from importlib import metadata