+++ examples/502_dag.py
@@ -16,7 +16,7 @@ import plotly
 from elphick.mass_composition import MassComposition
 from elphick.mass_composition.dag import DAG
 from elphick.mass_composition.datasets.sample_data import sample_data
-from elphick.mass_composition.network import MCNetwork
+from elphick.mass_composition.flowsheet import Flowsheet
 
 # %%
 # Define the DAG
@@ -48,14 +48,14 @@ dag.run({'feed_1': mc_sample,
          })
 
 # %%
-# Create a MCNetwork object from the dag, enabling all the usual network plotting and analysis methods.
+# Create a Flowsheet object from the dag, enabling all the usual network plotting and analysis methods.
 
-mcn: MCNetwork = MCNetwork.from_dag(dag)
+fs: Flowsheet = Flowsheet.from_dag(dag)
 
-fig = mcn.plot_network()
+fig = fs.plot_network()
 plotly.io.show(fig)
 
 # %%
 
-fig = mcn.table_plot(plot_type='sankey')
+fig = fs.table_plot(plot_type='sankey')
 fig